// app/types/member.ts
export interface Member {
  id: string;
  birthDate: string;
  bloodType: string;
  breakfastFrequency: string;
  city: string;
  createdAt: { seconds: number; nanoseconds: number };
  eatingReasons: string[];
  email: string;
  emergencyName: string;
  emergencyPhone: string;
  exerciseDays: string[];
  exerciseDuration: string;
  exercisePain: boolean;
  exerciseTime: string;
  firstName: string;
  foodTracking: boolean;
  goalWeight: string;
  healthIssues: string;
  height: string;
  lastName: string;
  medications: string;
  membershipType: string;
  nightEating: string;
  nutritionRating: string;
  password: string;
  phoneNumber: string;
  proSport: boolean;
  role: string;
  signature: string;
  smoke: boolean;
  startMonth: string;
  state: string;
  streetAddress: string;
  streetAddress2: string;
  supplements: boolean;
  surgery: boolean;
  trainingGoals: string[];
  userId: string;
  weight: string;
  zipCode: string;
  collectionType: string;
  name?: string;
  membership?: string;
  status?: string;
  statusColor?: string;
  [key: string]: any;
}